python-service/requirements.txt
flask==3.0.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
python-dotenv==1.0.0
psycopg2==2.9.9
requests==2.31.0
pydantic==2.5.2
python-dateutil==2.8.2
pytz==2023.3
